requirements.txt
ollama
PyYAML
requests